added type helper
