+++ src/gen/types.rs
@@ -1,5 +1,293 @@
+use std::{cell::RefCell, collections::{HashSet, BTreeSet}};
+
 use genco::prelude::*;
-use uniffi_bindgen::interface::{FfiType, Type};
+use uniffi_bindgen::{interface::{FfiType, Type}, ComponentInterface};
+
+use super::{Config, render::{Renderable, Renderer}};
+use crate::gen::primitives;
+
+
+#[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
+pub enum ImportRequirement {
+    Import { name: String },
+    ImportAs { name: String, as_name: String },
+}
+
+pub struct TypeHelpersRenderer<'a> {
+    config: &'a Config,
+    ci: &'a ComponentInterface,
+    include_once_names: RefCell<HashSet<String>>,
+    imports: RefCell<BTreeSet<ImportRequirement>>,
+}
+
+impl<'a> TypeHelpersRenderer<'a> {
+    pub fn new(config: &'a Config, ci: &'a ComponentInterface) -> Self {
+        Self {
+            config,
+            ci,
+            include_once_names: RefCell::new(HashSet::new()),
+            imports: RefCell::new(BTreeSet::new()),
+        }
+    }
+
+    fn external_type_package_name(&self, crate_name: &str) -> String {
+        match self.config.external_packages.get(crate_name) {
+            Some(name) => name.clone(),
+            None => crate_name.to_string(),
+        }
+    }
+
+    fn include_once_check(&self, name: &str) -> bool {
+        self.include_once_names
+            .borrow_mut()
+            .insert(name.to_string())
+    }
+
+    fn add_import(&self, name: &str) -> &str {
+        self.imports.borrow_mut().insert(ImportRequirement::Import {
+            name: name.to_owned(),
+        });
+        ""
+    }
+
+    fn add_import_as(&self, name: &str, as_name: &str) -> &str {
+        self.imports
+            .borrow_mut()
+            .insert(ImportRequirement::ImportAs {
+                name: name.to_owned(),
+                as_name: as_name.to_owned(),
+            });
+        ""
+    }
+}
+
+impl Renderable for TypeHelpersRenderer<'_> {
+    fn render_type(&self, ty: &uniffi_bindgen::interface::Type) -> dart::Tokens {
+        todo!()
+    }
+
+    fn render_type_helpers(&self, ty: &uniffi_bindgen::interface::Type) -> dart::Tokens {
+        todo!()
+    }
+}
+
+impl Renderer for TypeHelpersRenderer<'_> {
+    fn render(&self, r: &impl Renderable) -> dart::Tokens {
+        quote! {
+            import "dart:async";
+            import "dart:convert";
+            import "dart:ffi";
+            import "dart:io" show Platform, File, Directory;
+            import "dart:isolate";
+            import "dart:typed_data";
+            import "package:ffi/ffi.dart";
+
+
+            class UniffiInternalError implements Exception {
+                static const int bufferOverflow = 0;
+                static const int incompleteData = 1;
+                static const int unexpectedOptionalTag = 2;
+                static const int unexpectedEnumCase = 3;
+                static const int unexpectedNullPointer = 4;
+                static const int unexpectedRustCallStatusCode = 5;
+                static const int unexpectedRustCallError = 6;
+                static const int unexpectedStaleHandle = 7;
+                static const int rustPanic = 8;
+
+                final int errorCode;
+                final String? panicMessage;
+
+                const UniffiInternalError(this.errorCode, this.panicMessage);
+
+                static UniffiInternalError panicked(String message) {
+                return UniffiInternalError(rustPanic, message);
+                }
+
+                @override
+                String toString() {
+                switch (errorCode) {
+                    case bufferOverflow:
+                    return "UniFfi::BufferOverflow";
+                    case incompleteData:
+                    return "UniFfi::IncompleteData";
+                    case unexpectedOptionalTag:
+                    return "UniFfi::UnexpectedOptionalTag";
+                    case unexpectedEnumCase:
+                    return "UniFfi::UnexpectedEnumCase";
+                    case unexpectedNullPointer:
+                    return "UniFfi::UnexpectedNullPointer";
+                    case unexpectedRustCallStatusCode:
+                    return "UniFfi::UnexpectedRustCallStatusCode";
+                    case unexpectedRustCallError:
+                    return "UniFfi::UnexpectedRustCallError";
+                    case unexpectedStaleHandle:
+                    return "UniFfi::UnexpectedStaleHandle";
+                    case rustPanic:
+                    return "UniFfi::rustPanic: $$panicMessage";
+                    default:
+                    return "UniFfi::UnknownError: $$errorCode";
+                }
+                }
+            }
+
+            const int CALL_SUCCESS = 0;
+            const int CALL_ERROR = 1;
+            const int CALL_PANIC = 2;
+
+            class RustCallStatus extends Struct {
+                @Int8()
+                external int code;
+                external RustBuffer errorBuf;
+
+                static Pointer<RustCallStatus> allocate({int count = 1}) =>
+                calloc<RustCallStatus>(count * sizeOf<RustCallStatus>()).cast();
+            }
+
+            T noop<T>(T t) {
+                return t;
+            }
+
+            T rustCall<T>(Api api, T Function(Pointer<RustCallStatus>) callback) {
+                var callStatus = RustCallStatus.allocate();
+                final returnValue = callback(callStatus);
+
+                switch (callStatus.ref.code) {
+                case CALL_SUCCESS:
+                    calloc.free(callStatus);
+                    return returnValue;
+                case CALL_ERROR:
+                    throw callStatus.ref.errorBuf;
+                case CALL_PANIC:
+                    if (callStatus.ref.errorBuf.len > 0) {
+                        final message = liftString(api, callStatus.ref.errorBuf.toIntList());
+                        calloc.free(callStatus);
+                        throw UniffiInternalError.panicked(message);
+                    } else {
+                        calloc.free(callStatus);
+                        throw UniffiInternalError.panicked("Rust panic");
+                    }
+                default:
+                    throw UniffiInternalError(callStatus.ref.code, null);
+                }
+            }
+
+            class RustBuffer extends Struct {
+                @Int32()
+                external int capacity;
+
+                @Int32()
+                external int len;
+
+                external Pointer<Uint8> data;
+
+                static RustBuffer fromBytes(Api api, ForeignBytes bytes) {
+                    final _fromBytesPtr = api._lookup<
+                    NativeFunction<
+                        RustBuffer Function(ForeignBytes, Pointer<RustCallStatus>)>>($(format!("\"{}\"", self.ci.ffi_rustbuffer_from_bytes().name())));
+                    final fromBytes =
+                    _fromBytesPtr.asFunction<RustBuffer Function(ForeignBytes, Pointer<RustCallStatus>)>();
+                    return rustCall(api, (res) => fromBytes(bytes, res));
+                }
+
+                void deallocate(Api api) {
+                    final _freePtr = api._lookup<
+                    NativeFunction<
+                        Void Function(RustBuffer, Pointer<RustCallStatus>)>>($(format!("\"{}\"", self.ci.ffi_rustbuffer_free().name())));
+                    final free = _freePtr.asFunction<void Function(RustBuffer, Pointer<RustCallStatus>)>();
+                    rustCall(api, (res) => free(this, res));
+                }
+
+                Uint8List toIntList() {
+                    final buf = Uint8List(len);
+                    final precast = data.cast<Uint8>();
+                    for (int i = 0; i < len; i++) {
+                        buf[i] = precast.elementAt(i).value;
+                    }
+                    return buf;
+                }
+
+                @override
+                String toString() {
+                    String res = "RustBuffer { capacity: $capacity, len: $len, data: $data }";
+                    final precast = data.cast<Uint8>();
+                    for (int i = 0; i < len; i++) {
+                        int char = precast.elementAt(i).value;
+                        res += String.fromCharCode(char);
+                    }
+                    return res;
+                }
+            }
+
+            String liftString(Api api, Uint8List input) {        
+                // we have a i32 length at the front
+                return utf8.decoder.convert(input);
+            }
+
+            $(primitives::generate_primitives_lifters())
+           
+            Uint8List lowerString(Api api, String input) {
+                // FIXME: this is too many memcopies!
+                return Utf8Encoder().convert(input);
+            }
+
+            $(primitives::generate_primitives_lowerers())
+
+            RustBuffer toRustBuffer(Api api, Uint8List data) {
+                final length = data.length;
+
+                final Pointer<Uint8> frameData = calloc<Uint8>(length); // Allocate a pointer large enough.
+                final pointerList = frameData.asTypedList(length); // Create a list that uses our pointer and copy in the data.
+                pointerList.setAll(0, data); // FIXME: can we remove this memcopy somehow?
+
+                final bytes = calloc<ForeignBytes>();
+                bytes.ref.len = length;
+                bytes.ref.data = frameData;
+                return RustBuffer.fromBytes(api, bytes.ref);
+            }
+
+            T? liftOptional<T>(Api api, Uint8List buf, T? Function(Api, Uint8List) lifter) {
+                if (buf.isEmpty || buf.first == 0){
+                    return null;
+                }
+                return lifter(api, buf);
+            }
+
+            $(primitives::generate_wrapper_lifters())
+
+            Uint8List lowerOptional<T>(Api api, T? inp, Uint8List Function(Api, T) lowerer) {
+                if (inp == null) {
+                    final res = Uint8List(1);
+                    res.first = 0;
+                    return res;
+                }
+                // converting the inner
+                final inner = lowerer(api, inp);
+                // preparing the outer
+                final offset = 5;
+                final res = Uint8List(inner.length + offset);
+                // first byte sets the option to as true
+                res.setAll(0, [1]);
+                // then set the inner size
+                final len = Uint32List(1);
+                len.first = inner.length;
+                res.setAll(1, len.buffer.asUint8List().reversed);
+                // then add the actual data
+                res.setAll(offset, inner);
+                return res;
+            }
+
+            $(primitives::generate_wrapper_lowerers())
+
+            class ForeignBytes extends Struct {
+                @Int32()
+                external int len;
+
+                external Pointer<Uint8> data;
+            }
+        }
+    }
+}
+
 
 pub fn generate_ffi_type(ret: Option<&FfiType>) -> dart::Tokens {
     let Some(ret_type) = ret else {
